improvements from review
